# src/main/resources/desafio1/api-docs.json
{"openapi":"3.0.1","info":{"title":"API Prospect Cielo","description":"API do desafio bootcamp Cielo promovido pela Ada Tech cujo objetivo e prospectar novos clientes.","contact":{"name":"Marcos Sousa","email":"backend@adatech.com"},"license":{"name":"Apache 2.0","url":"http://empresa/api/licenca"}},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/clientes/pessoasjuridicas":{"get":{"tags":["/clientes/pessoasjuridicas"],"summary":"Lista todas as pessoas jurídicas cadastradas","operationId":"listarPessoas","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ListagemPessoaJuridica"}}}}}}},"put":{"tags":["/clientes/pessoasjuridicas"],"summary":"Atualiza os dados de uma pessoa jurídica","operationId":"atualizar","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CadastroPessoaJuridica"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ListagemPessoaJuridica"}}}}}},"post":{"tags":["/clientes/pessoasjuridicas"],"summary":"Realiza o cadastro de uma pessoa jurídica","operationId":"cadastrar","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CadastroPessoaJuridica"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/clientes/pessoasfisicas":{"get":{"tags":["/clientes/pessoasfisicas"],"summary":"Lista todas as pessoas físicas cadastradas","operationId":"listarPessoas_1","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ListagemPessoaFisica"}}}}}}},"put":{"tags":["/clientes/pessoasfisicas"],"summary":"Atualiza os dados de uma pessoa física","operationId":"atualizar_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CadastroPessoaFisica"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ListagemPessoaFisica"}}}}}},"post":{"tags":["/clientes/pessoasfisicas"],"summary":"Realiza o cadastro de uma pessoa física","operationId":"cadastrar_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CadastroPessoaFisica"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/clientes/pessoasjuridicas/{uuid}":{"get":{"tags":["/clientes/pessoasjuridicas"],"summary":"Detalha os dados de uma pessoa jurídica cadastrada","operationId":"detalharPessoa","parameters":[{"name":"uuid","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}},"delete":{"tags":["/clientes/pessoasjuridicas"],"summary":"Realiza a exclusão de uma pessoa jurídica cadastrada","operationId":"excluir","parameters":[{"name":"uuid","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/clientes/pessoasfisicas/{uuid}":{"get":{"tags":["/clientes/pessoasfisicas"],"summary":"Detalha os dados de uma pessoa física cadastrada","operationId":"detalharPessoa_1","parameters":[{"name":"uuid","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}},"delete":{"tags":["/clientes/pessoasfisicas"],"summary":"Realiza a exclusão de uma pessoa física cadastrada","operationId":"excluir_1","parameters":[{"name":"uuid","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}}},"components":{"schemas":{"CadastroPessoaJuridica":{"required":["cnpj","cpf","emailContato","mcc","nomeContato","razaoSocial"],"type":"object","properties":{"mcc":{"pattern":"\\d{4}","type":"string"},"cpf":{"maxLength":11,"minLength":0,"type":"string"},"cnpj":{"maxLength":14,"minLength":0,"type":"string"},"razaoSocial":{"maxLength":50,"minLength":0,"type":"string"},"nomeContato":{"maxLength":50,"minLength":0,"type":"string"},"emailContato":{"type":"string"}}},"ListagemPessoaJuridica":{"type":"object","properties":{"uuid":{"type":"string","format":"uuid"},"mcc":{"type":"string"},"razaoSocial":{"type":"string"},"cnpj":{"type":"string"},"nomeContato":{"type":"string"},"emailContato":{"type":"string"}}},"CadastroPessoaFisica":{"required":["cpf","emailPessoa","mcc","nomePessoa"],"type":"object","properties":{"mcc":{"pattern":"\\d{4}","type":"string"},"cpf":{"maxLength":11,"minLength":0,"type":"string"},"nomePessoa":{"maxLength":50,"minLength":0,"type":"string"},"emailPessoa":{"type":"string"}}},"ListagemPessoaFisica":{"type":"object","properties":{"uuid":{"type":"string","format":"uuid"},"mcc":{"type":"string"},"cpf":{"type":"string"},"nomePessoa":{"type":"string"},"emailPessoa":{"type":"string"}}}}}}